+++ FERRETERIA  Joel/Models/Proveedor.cs
@@ -2,7 +2,7 @@
 {
     public class Proveedor
     {
-        public short IdProveedor { get; set; }
+        public int IdProveedor { get; set; }
         public string RazonSocial { get; set; } = string.Empty;
         public string NombreComercial { get; set; } = string.Empty;
         public string Nit { get; set; } = string.Empty;
@@ -13,6 +13,6 @@
         public byte Estado { get; set; }
         public DateTime FechaRegistro { get; set; }
         public DateTime? FechaActualizacion { get; set; }
-        public short IdEmpleadoResponsable { get; set; }
+        public int IdEmpleadoResponsable { get; set; }
     }
 }
